pyproject.toml
[tool.poetry]
name = "a-flask-test"
version = "0.1.0"
description = ""
authors = ["rochacbruno"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
accelerate = "1.1.1"
altair = "5.4.1"
annotated-types = "0.7.0"
anyio = "4.5.2"
asgiref = "3.8.1"
attrs = "24.2.0"
backoff = "2.2.1"
bcrypt = "4.2.0"
black = "24.10.0"
blinker = "1.9.0"
bson = "0.5.10"
build = "1.2.2.post1"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
chroma-hnswlib = "0.7.6"
chromadb = "0.5.18"
click = "8.1.7"
coloredlogs = "15.0.1"
colpali-engine = "0.3.4"
coverage = "7.6.1"
deprecated = "1.2.14"
distro = "1.9.0"
dnspython = "2.6.1"
durationpy = "0.9"
einops = "0.8.0"
fastapi = "0.115.4"
filelock = "3.16.1"
flake8 = "7.1.1"
flatbuffers = "24.3.25"
fsspec = "2024.10.0"
ghp-import = "2.1.0"
gitchangelog = "3.0.4"
gitdb = "4.0.11"
gitpython = "3.1.43"
google-auth = "2.35.0"
googleapis-common-protos = "1.65.0"
gputil = "1.4.0"
grpcio = "1.67.1"
h11 = "0.14.0"
httpcore = "1.0.6"
httptools = "0.6.4"
httpx = "0.27.2"
huggingface-hub = "0.26.2"
humanfriendly = "10.0"
idna = "3.10"
importlib-metadata = "8.4.0"
importlib-resources = "6.4.5"
iniconfig = "2.0.0"
isort = "5.13.2"
jinja2 = "3.1.4"
jiter = "0.6.1"
joblib = "1.4.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
kubernetes = "31.0.0"
markdown = "3.7"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mccabe = "0.7.0"
mdurl = "0.1.2"
mergedeep = "1.3.4"
mkdocs = "1.6.1"
mkdocs-get-deps = "0.2.0"
mmh3 = "5.0.1"
monotonic = "1.6"
mpmath = "1.3.0"
mypy = "1.13.0"
mypy-extensions = "1.0.0"
narwhals = "1.13.5"
networkx = "3.2.1"
numpy = "1.26.4"
oauthlib = "3.2.2"
ollama = "0.3.3"
onnxruntime="1.16.3"
openai = "1.51.2"
opencv-python = "4.10.0.84"
opentelemetry-api = "1.27.0"
opentelemetry-exporter-otlp-proto-common = "1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "1.27.0"
opentelemetry-instrumentation = "0.48b0"
opentelemetry-instrumentation-asgi = "0.48b0"
opentelemetry-instrumentation-fastapi = "0.48b0"
opentelemetry-proto = "1.27.0"
opentelemetry-sdk = "1.27.0"
opentelemetry-semantic-conventions = "0.48b0"
opentelemetry-util-http = "0.48b0"
orjson = "3.10.10"
overrides = "7.7.0"
packaging = "24.1"
pandas = "2.2.3"
pathspec = "0.12.1"
peft = "0.11.1"
pillow = "10.4.0"
platformdirs = "4.3.6"
pluggy = "1.5.0"
posthog = "3.7.0"
protobuf = "4.25.5"
psutil = "6.1.0"
pyarrow = ">=17.0.0,<17.1.0"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pycodestyle = "2.12.1"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pydeck = "0.9.1"
pyflakes = "3.2.0"
pygments = "2.18.0"
pymongo = "4.10.1"
pypika = "0.48.9"
pyproject-hooks = "1.2.0"
pytest = "8.3.3"
pytest-cov = "5.0.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
pyyaml = "6.0.2"
pyyaml-env-tag = "0.1"
referencing = "0.35.1"
regex = "2024.11.6"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.9.3"
rpds-py = "0.21.0"
rsa = "4.9"
safetensors = "0.4.5"
scikit-learn = "1.5.2"
scipy = "1.14.1"
sentence-transformers = "3.3.0"
setuptools = ">=75.2.0,<75.3.0"
shellingham = "1.5.4"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
starlette = "0.41.2"
streamlit = "1.40.1"
streamlit-vis-timeline = "0.3.0"
sympy = "1.13.1"
tenacity = "9.0.0"
threadpoolctl = "3.5.0"
timm = "1.0.11"
tokenizers = "0.20.3"
toml = "0.10.2"
torch = "2.2.2"
torchvision = "0.17.2"
tornado = "6.4.1"
tqdm = "4.67.1"
transformers = "4.46.2"
typer = "0.12.5"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
uvicorn = "0.32.0"
uvloop = "0.21.0"
watchdog = "4.0.2"
watchfiles = "0.24.0"
websocket-client = "1.8.0"
websockets = "13.1"
wheel = "0.44.0"
wrapt = "1.16.0"
xxhash = "3.5.0"
zipp = "3.20.2"
ijson = "^3.3.0"
infinity-embedded-sdk = "0.5.0.dev5"
pysqlite3 = "^0.5.4"
tortoise-orm = "^0.22.1"
pytest-asyncio = "^0.24.0"
google-search-results = "2.4.2"
serpapi = "0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aiki = 'aiki.__main__:main'
